Extract `stopwatch` to its own package
