+++ run.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/Shresht7/go-time/cmd"
 	"github.com/Shresht7/go-time/tui"
+	"github.com/Shresht7/go-time/tui/stopwatch"
 )
 
 // ---
@@ -43,7 +44,7 @@ func run(args []string, stdout io.Writer) error {
 		cmd.ShowCalendar(now)
 
 	case "stopwatch":
-		tui.ShowStopwatch()
+		stopwatch.Run()
 
 	default:
 		cmd.ShowTime(now)
